Penambahan SweetAlert pada admin

+++ admin/index.php
@@ -3,28 +3,38 @@
 <head>
 	<title>Membuat CRUD Dengan PHP Dan MySQL</title>
 	<link rel="stylesheet" type="text/css" href="style.css">
+	<script src="https://unpkg.com/sweetalert/dist/sweetalert.min.js"></script>
+
 </head>
 <body>
-	<div class="judul">		
+	<div class="judul">
 		<h1>Membuat CRUD Dengan PHP Dan MySQL</h1>
 		<h2>Menampilkan data dari database</h2>
 	</div>
 	<br/>
-	<?php 
-		if(isset($_GET['pesan'])){
-			$pesan = $_GET['pesan'];
-			if($pesan == "input"){
-				echo "Data berhasil di input.";
-			}else if($pesan == "update"){
-				echo "Data berhasil di update.";
-			}else if($pesan == "hapus"){
-				echo "Data berhasil di hapus.";
-			}elseif ($pesan == "gagal") {
-				echo "Data gagal di input . pastikan anda menulis form dengan benar";
-			}
-		}
+	<?php
+		// if(isset($_GET['pesan'])){
+		// 	$pesan = $_GET['pesan'];
+		// 	if($pesan == "input"){
+		// 		echo '<script type="text/javascript"">
+		// 		swal("Data", "Berhasil di input", "success");
+		// 		</script>';
+		// 	}else if($pesan == "update"){
+		// 		echo '<script type="text/javascript"">
+ 		// 	 swal("Data", "Berhasil di update", "success");
+ 		// 	 </script>';
+		// 	}else if($pesan == "hapus"){
+		// 		echo '<script type="text/javascript"">
+ 		// 	 swal("Data", "Berhasil di hapus", "success");
+ 		// 	 </script>';
+		// 	}elseif ($pesan == "gagal") {
+		// 		echo '<script type="text/javascript"">
+ 		// 	 swal("Data", "Gagal di input", "success");
+ 		// 	 </script>';
+		// 	}
+		//}
 		?>
-		
+
 		<br/>
 		<br/>
 		<form action="index.php" method="get">
@@ -32,23 +42,33 @@
 		<input type="text" name="cari">
 		<input type="submit" value="Cari">
 		</form>
-		<?php 
+		<?php
 		if(isset($_GET['cari'])){
 			$cari = $_GET['cari'];
 			echo "<b>Hasil pencarian : ".$cari."</b>";
 		}
 		?>
 		<br/>
-		
-	<?php 
+
+	<?php
 	if(isset($_GET['pesan'])){
 		$pesan = $_GET['pesan'];
 		if($pesan == "input"){
-			echo "Data berhasil di input.";
+			echo '<script type="text/javascript"">
+			swal("Data", "Berhasil di input", "success");
+			</script>';
 		}else if($pesan == "update"){
-			echo "Data berhasil di update.";
+			 echo '<script type="text/javascript"">
+			 swal("Data", "Berhasil di update", "success");
+			 </script>';
 		}else if($pesan == "hapus"){
-			echo "Data berhasil di hapus.";
+			echo '<script type="text/javascript"">
+			swal("Data", "Berhasil di hapus", "success");
+			</script>';
+		}elseif ($pesan == "gagal") {
+			echo '<script type="text/javascript"">
+		 swal("Data", "Gagal di proses", "error");
+		 </script>';
 		}
 	}
 	?>
@@ -64,22 +84,22 @@
 			<th>Pekerjaan</th>
 			<th>Email</th>
 			<th>No. Telpon</th>
-			<th>Opsi</th>		
+			<th>Opsi</th>
 		</tr>
-		
+
 		<?php
 		include "koneksi.php";
-		
+
 		if(isset($_GET['cari'])){
 			$cari = $_GET['cari'];
-			$query_mysql = mysql_query("select * from user where nama like '%".$cari."%'");				
+			$query_mysql = mysql_query("select * from user where nama like '%".$cari."%'");
 		}else{
-			$query_mysql = mysql_query("select * from user")or die(mysql_error());		
+			$query_mysql = mysql_query("select * from user")or die(mysql_error());
 		}
 		$nomor = 1;
 		while($data = mysql_fetch_array($query_mysql)){
 		?>
-		
+
 		<tr>
 			<td><?php echo $nomor++; ?></td>
 			<td><?php echo $data['nama']; ?></td>
@@ -89,15 +109,16 @@
 			<td><?php echo $data['no_telp']; ?></td>
 			<td>
 				<a href="edit.php?id=<?php echo $data['id']; ?>">Edit</a> |
-				<a href="hapus.php?id=<?php echo $data['id']; ?>">Hapus</a>					
+				<a href="hapus.php?id=<?php echo $data['id']; ?>" onclick="javascript: return confirm('Anda yakin hapus ?')">Hapus</a>
 			</td>
 		</tr>
 		<?php } ?>
 	</table>
+
 </body>
 </html>
 
-<?php 
+<?php
 //include '../config.php';
 
 // mengaktifkan session
